# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MovieHost API",
    "description": "Media library, playback progress, and viewer profiles.",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
    "/api/details": {
      "get": {
        "tags": [
          "media"
        ],
        "summary": "Title metadata for a media path.",
        "operationId": "handle_details",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "description": "Media path relative to the server's media root.",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "Movies/Arrival.mkv"
          }
        ],
        "responses": {
          "200": {
            "description": "Title metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieDetails"
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No metadata for this path",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/profile/{id}/watch_state": {
      "put": {
        "tags": [
          "profiles"
        ],
        "summary": "Record how far a profile got through one piece of media.",
        "operationId": "handle_put_watch_state",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Profile id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchState"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Watch state saved"
          },
          "400": {
            "description": "Malformed id or body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/thumbnail": {
      "get": {
        "tags": [
          "media"
        ],
        "summary": "Poster image for a media path, proxied and cached from TMDB.",
        "operationId": "handle_thumbnail",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "description": "Media path relative to the server's media root.",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "Movies/Arrival.mkv"
          }
        ],
        "responses": {
          "200": {
            "description": "Poster image",
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int32",
                    "minimum": 0
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing or malformed query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No artwork for this path",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "description": "JSON body returned for every error response.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable description of what went wrong.",
            "example": "Not found"
          }
        }
      },
...
      "MovieDetails": {
        "type": "object",
        "description": "Metadata for one title, as served by `/api/details`.",
        "properties": {
          "overview": {
            "type": [
              "string",
              "null"
            ]
          },
          "poster_path": {
            "type": [
              "string",
              "null"
            ],
            "description": "TMDB-relative poster path, absent when TMDB has no artwork."
          },
          "release_date": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "WatchState": {
        "type": "object",
        "description": "How far a profile got through one piece of media. Doubles as the request\nbody for `PUT /api/profile/{id}/watch_state`.",
        "required": [
          "movie_path",
          "last_position",
          "finished"
        ],
        "properties": {
          "finished": {
            "type": "boolean"
          },
          "last_position": {
            "type": "number",
            "format": "float",
            "description": "Playback position in seconds."
          },
          "movie_path": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}